backend/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn[standard]>=0.22.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
streamlit>=1.30.0
requests>=2.31.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
chromadb>=0.4.22
pypdf>=3.17.0
gunicorn>=21.2.0
uvloop>=0.19.0
